+++ core/src/elements/ons-navigator/index.js
@@ -363,13 +363,11 @@ class NavigatorElement extends BaseElement {
       options.animationOptions = util.extend({}, leavePage.pushedOptions.animationOptions, options.animationOptions || {});
 
       const callback = () => {
-        setImmediate(() => enterPage._show());
-        leavePage._hide();
-
         pages.pop();
         update(pages, this).then(() => {
           this._isRunning = false;
 
+          enterPage._show();
           util.triggerElementEvent(this, 'postpop', {leavePage, enterPage, navigator: this});
 
           if (typeof options.callback === 'function') {
@@ -380,6 +378,7 @@ class NavigatorElement extends BaseElement {
         });
       };
 
+      leavePage._hide();
       const animator = this._animatorFactory.newAnimator(options);
       animator.pop(this.pages[l - 2], this.pages[l - 1], callback);
     }).catch(() => this._isRunning = false);
@@ -460,6 +459,11 @@ class NavigatorElement extends BaseElement {
 
       var enterPage  = this.pages[pageLength - 1];
       var leavePage = this.pages[pageLength - 2];
+
+      if (enterPage.nodeName !== 'ONS-PAGE') {
+        throw new Error('Only elements of type <ons-page> can be pushed to the navigator');
+      }
+
       enterPage.updateBackButton(pageLength - 1);
 
       enterPage.data = options.data;
@@ -474,6 +478,7 @@ class NavigatorElement extends BaseElement {
             leavePage.style.display = 'none';
           }
 
+          enterPage._show();
           util.triggerElementEvent(this, 'postpush', {leavePage, enterPage, navigator: this});
 
           if (typeof options.callback === 'function') {
@@ -489,17 +494,18 @@ class NavigatorElement extends BaseElement {
           enterPage.style.display = 'block';
           if (leavePage) {
             leavePage._hide();
-            setImmediate(() => enterPage._show());
             animator.push(enterPage, leavePage, done);
           } else {
-            setImmediate(() => enterPage._show());
             done();
           }
         };
 
         options._linked ? push() : rewritables.link(this, enterPage, options, push);
       });
-    }).catch(() => this._isRunning = false);
+    }).catch((error) => {
+      this._isRunning = false;
+      throw error;
+    });
   }
 
   /**